xml_export script: try to fix the '[: too many arguments' error intermittently seen when running this cron script.
It's suspected that the above error is a result of bash tokenising the input. The key change here is wrapping the $(command substitution) in double-quotes, so that input is treated as one item.
There is also another change to take the 'cd' command out of the loop, as this appears to be unnecessary and not clear why the original script's author included it in the loop.
See also Salesforce case 00487508.